rag_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
qdrant-client==1.12.1
sentence-transformers==3.3.0
asyncpg==0.30.0
google-generativeai==0.8.4
python-dotenv==1.0.1
pydantic==2.10.1
python-multipart==0.0.20